Rename top-level files to reduce unintended import issues
