+++ sbot/_comp.py
