+++ sbot/_vision.py
